refactor: refactor swagger-ui module with TS
